Interaction System
Okay for soem reason Unity is being shit weird, and I can't click on random things and its just terrible.

So I made the pseudocode for the interaction system. Basically anything that can be interacted with when you get close enough and press E will have the Interactable script on it, then the code will doTheThing() when you press E. It should be pretty easy to fill the things in the inspector and actually writ ethe real code as I'm pretty sure the system should work.